+++ croud/__main__.py
@@ -49,12 +49,14 @@ from croud.cmd import (
     resource_id_arg,
     role_fqn_arg,
     user_id_arg,
+    user_id_or_email_arg,
 )
 from croud.config import Configuration, config_get, config_set
 from croud.login import login
 from croud.logout import logout
 from croud.me import me
 from croud.organizations.commands import organizations_create, organizations_list
+from croud.organizations.users.commands import org_users_add, org_users_remove
 from croud.products.deploy import product_deploy
 from croud.projects.commands import project_create, projects_list
 from croud.users.commands import users_list
@@ -159,6 +161,36 @@ def main():
                     "extra_args": [output_fmt_arg],
                     "calls": organizations_list,
                 },
+                "users": {
+                    "help": "Add/remove users to/from organizations.",
+                    "sub_commands": {
+                        "add": {
+                            "help": "Add user to organization",
+                            "extra_args": [
+                                user_id_or_email_arg,
+                                lambda req_opt_group, opt_opt_group: role_fqn_arg(
+                                    req_opt_group, opt_opt_group, False
+                                ),
+                                lambda req_opt_group, opt_opt_group: org_id_arg(
+                                    req_opt_group, opt_opt_group, False
+                                ),
+                            ],
+                            "calls": org_users_add,
+                        },
+                        "remove": {
+                            "help": "Remove user from organization",
+                            "extra_args": [
+                                lambda req_opt_group, opt_opt_group: user_id_arg(
+                                    req_opt_group, opt_opt_group, True
+                                ),
+                                lambda req_opt_group, opt_opt_group: org_id_arg(
+                                    req_opt_group, opt_opt_group, False
+                                ),
+                            ],
+                            "calls": org_users_remove,
+                        },
+                    },
+                },
             },
         },
         "users": {
@@ -189,7 +221,9 @@ def main():
                                     req_opt_group, opt_opt_group, True
                                 ),
                                 output_fmt_arg,
-                                role_fqn_arg,
+                                lambda req_opt_group, opt_opt_group: role_fqn_arg(
+                                    req_opt_group, opt_opt_group, True
+                                ),
                             ],
                             "calls": roles_add,
                         },
@@ -203,7 +237,9 @@ def main():
                                     req_opt_group, opt_opt_group, True
                                 ),
                                 output_fmt_arg,
-                                role_fqn_arg,
+                                lambda req_opt_group, opt_opt_group: role_fqn_arg(
+                                    req_opt_group, opt_opt_group, True
+                                ),
                             ],
                             "calls": roles_remove,
                         },